+++ packages/elements/mention/examples/08-resourced-mention-on-n20-background.tsx
@@ -1,16 +1,16 @@
 import * as React from 'react';
 import { AnalyticsListener } from '@atlaskit/analytics';
+import { colors } from '@atlaskit/theme';
 import debug from '../src/util/logger';
 import { onMentionEvent } from '../example-helpers/index';
 import {
   mockMentionData as mentionData,
   mockMentionProvider as mentionProvider,
 } from '../src/__tests__/unit/_test-helpers';
-import { akColorN20 } from '@atlaskit/util-shared-styles';
 import ResourcedMention from '../src/components/Mention/ResourcedMention';
 
 const style = {
-  backgroundColor: akColorN20,
+  backgroundColor: colors.N20,
   width: '100%',
   height: '800px',
 };