+++ packages/elements/user-picker/src/components/components.ts
@@ -1,21 +1,32 @@
+import { components } from '@atlaskit/select';
 import memoizeOne from 'memoize-one';
-import { UserMultiValueLabel } from './UserMultiValueLabel';
-import { UserMultiValueRemove } from './UserMultiValueRemove';
+import { ClearIndicator } from './ClearIndicator';
+import { MultiValue } from './MultiValue';
+import { Option } from './Option';
+import { Placeholder } from './Placeholder';
+import { SingleValue } from './SingleValue';
+import { ValueContainer } from './ValueContainer';
 
 /**
  * Memoize getComponents to avoid rerenders.
  */
-export const getComponents = memoizeOne((anchor?: React.ComponentType<any>) => {
-  if (anchor) {
-    return {
-      MultiValueLabel: UserMultiValueLabel,
-      MultiValueRemove: UserMultiValueRemove,
-      Control: anchor,
-    };
-  } else {
-    return {
-      MultiValueLabel: UserMultiValueLabel,
-      MultiValueRemove: UserMultiValueRemove,
-    };
-  }
-});
+export const getComponents = memoizeOne(
+  (multi?: boolean, anchor?: React.ComponentType<any>) => {
+    if (anchor) {
+      return {
+        Control: anchor,
+        Option: Option,
+      };
+    } else {
+      return {
+        MultiValue,
+        DropdownIndicator: null,
+        SingleValue,
+        ClearIndicator: multi ? null : ClearIndicator,
+        Placeholder,
+        Option: Option,
+        ValueContainer: multi ? ValueContainer : components.ValueContainer,
+      };
+    }
+  },
+);