+++ packages/elements/emoji/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/emoji
 
+## 45.0.3
+- [patch] [527b954](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/527b954):
+
+  - FS-3174 - Remove usage of util-shared-styles from elements components
+
 ## 45.0.2
 - [patch] Tweak emoji tests to work better with newest EmojiOne v4 metadata [c034007](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c034007)